model_servers/image_captioning_server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
Pillow==11.1.0
transformers[torch]==4.51.3
torch==2.7.0
sentencepiece==0.2.0

